[#165] Refactor: 냉장고 삭제 기능 멱등성 보장하도록 수정

+++ src/main/java/soon/fridgely/domain/refrigerator/service/RefrigeratorManager.java
@@ -54,8 +54,11 @@ public class RefrigeratorManager {
     @CacheEvict(value = "myRefrigerators", allEntries = true)
     @Transactional
     public void delete(long refrigeratorId) {
-        Refrigerator refrigerator = refrigeratorRepository.findByIdAndStatus(refrigeratorId, EntityStatus.ACTIVE)
+        Refrigerator refrigerator = refrigeratorRepository.findById(refrigeratorId)
             .orElseThrow(() -> new CoreException(ErrorType.NOT_FOUND_DATA));
+        if (refrigerator.isDeleted()) {
+            return;
+        }
         refrigerator.delete();
     }
